working

+++ src/gpu/AllReduceHIERSwingMPI.cu
@@ -1164,7 +1164,7 @@ int main(int argc, char *argv[]) {
     int *d_test_recv_buffer;
     CUDA_CHECK(cudaMalloc((void**)&d_test_recv_buffer, (size_t) BUFFER_SIZE));
     srand(time(NULL)*rank);      
-    for (int i = 0; i < msg_count; i++) {
+    for (size_t i = 0; i < msg_count; i++) {
         h_send_buffer[i] = rand()*rank % 10; 
     }
